requirements.txt
python-dotenv
selenium
pandas
openpyxl
jinja2
watchdog
tabulate
beautifulsoup4
